chore: push inlining info code into a submodule
