+++ compiler/noirc_evaluator/src/ssa/opt/preprocess_fns.rs
@@ -11,10 +11,10 @@ impl Ssa {
     /// Run pre-processing steps on functions in isolation.
     pub(crate) fn preprocess_functions(mut self, aggressiveness: i64) -> Ssa {
         // Bottom-up order, starting with the "leaf" functions, so we inline already optimized code into the ones that call them.
-        let bottom_up = inlining::compute_bottom_up_order(&self);
+        let bottom_up = inlining::inline_info::compute_bottom_up_order(&self);
 
         // Preliminary inlining decisions.
-        let inline_infos = inlining::compute_inline_infos(&self, false, aggressiveness);
+        let inline_infos = inlining::inline_info::compute_inline_infos(&self, false, aggressiveness);
 
         let should_inline_call = |callee: &Function| -> bool {
             match callee.runtime() {
